projects/motivational_quote_writer/requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
attrs==23.2.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
crewai==0.19.0
dataclasses-json==0.6.4
Deprecated==1.2.14
distro==1.9.0
docstring-parser==0.15
frozenlist==1.4.1
google-ai-generativelanguage==0.4.0
google-api-core==2.17.1
google-auth==2.28.2
google-generativeai==0.3.2
googleapis-common-protos==1.62.0
greenlet==3.0.3
grpcio==1.62.1
grpcio-status==1.62.1
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
idna==3.6
importlib-metadata==6.11.0
instructor==0.5.2
jsonpatch==1.33
jsonpointer==2.4
langchain==0.1.11
langchain-community==0.0.27
langchain-core==0.1.30
langchain-google-genai==0.0.9
langchain-openai==0.0.5
langchain-text-splitters==0.0.1
langsmith==0.1.23
markdown-it-py==3.0.0
marshmallow==3.21.1
mdurl==0.1.2
multidict==6.0.5
mypy-extensions==1.0.0
numpy==1.26.4
openai==1.13.3
opentelemetry-api==1.23.0
opentelemetry-exporter-otlp-proto-common==1.23.0
opentelemetry-exporter-otlp-proto-http==1.23.0
opentelemetry-proto==1.23.0
opentelemetry-sdk==1.23.0
opentelemetry-semantic-conventions==0.44b0
orjson==3.9.15
packaging==23.2
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==2.6.3
pydantic_core==2.16.3
Pygments==2.17.2
python-dotenv==1.0.1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
rich==13.7.1
rsa==4.9
-e git+https://github.com/andersonbosa/playground-crewai.git@f3d4d9b1efa0b51fa2b40d9675eeb799e5fad1e4#egg=shared_playground_crewai&subdirectory=projects/shared
sniffio==1.3.1
SQLAlchemy==2.0.28
tenacity==8.2.3
tiktoken==0.5.2
tqdm==4.66.2
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.10.0
urllib3==2.2.1
wrapt==1.16.0
yarl==1.9.4
zipp==3.17.0
